gatereader_requirements.txt
flask
Pillow
numpy
flask-cors
opencv-python
hailo-platform
